+++ src/app/center-bottom-content/center-bottom-content.component.spec.ts
@@ -35,7 +35,7 @@ describe('CenterBottomContentComponent', () => {
 
   beforeEach(() => {
     mockWorkflowService = {
-      selectedWorkflow: new Observable<Workflow[]>()
+      selectedWorkflow: new Observable<Workflow>()
     };
     fixture = TestBed.createComponent(CenterBottomContentComponent);
     workflowService = TestBed.inject(WorkflowService)
@@ -46,16 +46,16 @@ describe('CenterBottomContentComponent', () => {
 
   // Testing service dependencies 
   it('#selectedWorkflow should display a selected workflow values', () => {
-    const workflow: Workflow[] = [{
+    const workflow: Workflow = {
       title: "Delineation",
       description: "string",
       functionality: "string",
       icon: "string",
       steps: [],
       output: []
-    }];
-    component.selectedWorkflows = workflow;
-    expect(component.selectedWorkflows[0].title).toEqual(workflow[0].title);
+    };
+    component.selectedWorkflow = workflow;
+    expect(component.selectedWorkflow.title).toEqual(workflow.title);
   });
 
   afterEach(() => {